Add files via upload
The candy machine has four dispensers to hold and release items sold by the candy machine and a cash register. The machine sells four products — candies, chips, gum, and cookies — each of which is stored in a separate dispenser. You have been asked to write a program for this candy machine so that it can be put into operation